Initial pull apart of the backend into seperate parts
